# Laravel-capstone/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Video API",
        "description": "API for managing video resources",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "List all videos",
                "operationId": "91a3a5e7add32d9ee4fb5ad638e10d66",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Video"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "Create a video",
                "operationId": "de7b3eb3f2fce49c828dae36fff21116",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VideoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Video"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/videos/{id}": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get a video",
                "operationId": "039b82b6697daed032f32c1a5dfbb2d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Video"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Videos"
                ],
                "summary": "Update a video",
                "operationId": "b069962b1c673892a9bd6e130975f8fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VideoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Video updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Video"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Video not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Videos"
                ],
                "summary": "Delete a video",
                "operationId": "889cee044a8d2f628ee67ddb759c77d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Video deleted"
                    },
                    "404": {
                        "description": "Video not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Video": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Sample Video"
                    },
                    "description": {
                        "type": "string",
                        "example": "Video description"
                    },
                    "url": {
                        "type": "string",
                        "format": "url",
                        "example": "https://example.com/video.mp4"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "VideoInput": {
                "required": [
                    "title",
                    "url"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "New Video"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Video description"
                    },
                    "url": {
                        "type": "string",
                        "format": "url",
                        "maxLength": 255,
                        "example": "https://example.com/new-video.mp4"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "tags": [
        {
            "name": "Videos",
            "description": "All video endpoints"
        }
    ]
}